[FIX] test code 파일추가, url 변경 - CouponFeignController
